// src/app/api/models/workflow.ts
/* tslint:disable */
/* eslint-disable */
export interface Workflow {
  bankId: number;
  createdAt?: string;
  description?: null | string;
  id?: number;
  masterWorkflowId: number;
  status?: null | string;
  totalSteps: number;
  updatedAt?: string;
  version?: null | string;
  workflowCode?: null | string;
}